--- src/main/java/uk/gov/companieshouse/filing/processor/Unmarshaller.java
@@ -1,18 +1,0 @@
-package uk.gov.companieshouse.filing.processor;
-
-import java.io.IOException;
-
-import uk.gov.companieshouse.filing.model.Address;
-
-public interface Unmarshaller {
-    
-    /**
-     * Function to unmarshall the address from the data string
-     * 
-     * @param json
-     * @return
-     * @throws IOException
-     */
-    Address unmarshallAddress(String json) throws IOException;
-
-}
