--- src/main/java/uk/gov/companieshouse/filing/util/DateService.java
@@ -1,14 +1,0 @@
-package uk.gov.companieshouse.filing.util;
-
-import java.time.Instant;
-
-public interface DateService {
-
-    /**
-     * Obtains the current instant
-     * 
-     * @return the current instant, not null
-     */
-    Instant now();
-
-}
