--- src/main/java/uk/gov/companieshouse/filing/util/DateServiceImpl.java
@@ -1,14 +1,0 @@
-package uk.gov.companieshouse.filing.util;
-
-import java.time.Instant;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class DateServiceImpl implements DateService {
-
-    @Override
-    public Instant now() {
-        return Instant.now();
-    }
-}
