--- src/main/java/uk/gov/companieshouse/filing/writer/FilingWriter.java
@@ -1,15 +1,0 @@
-package uk.gov.companieshouse.filing.writer;
-
-import uk.gov.companieshouse.filing.model.FilingProcessed;
-
-@FunctionalInterface
-public interface FilingWriter {
-
-    /**
-     * Write a FilingProcessed to the kafka topic
-     * 
-     * @param filingProcessed The filing processed message
-     * @return
-     */
-    boolean write(FilingProcessed filingProcessed) throws FilingWriterException;
-}
