+++ src/main/java/uk/gov/companieshouse/filingmock/util/DateServiceImpl.java
@@ -1,0 +1,14 @@
+package uk.gov.companieshouse.filingmock.util;
+
+import java.time.Instant;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class DateServiceImpl implements DateService {
+
+    @Override
+    public Instant now() {
+        return Instant.now();
+    }
+}
