--- src/test/java/uk/gov/companieshouse/filing/writer/FilingWriterImplTest.java
@@ -1,18 +1,0 @@
-package uk.gov.companieshouse.filing.writer;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-@ExtendWith(MockitoExtension.class)
-public class FilingWriterImplTest {
-    
-    @InjectMocks
-    private FilingWriterImpl writer;
-    
-    @Test
-    public void write() throws Exception {
-        // TODO: implement unit test
-    }
-}
